# shelter-bak/src/swagger.yaml
swagger: '2.0'
info:
    description:
        This API provides main functionality to pets shelter
    version: 1.0.0
    title: Shelter Pets API
    contact:
        email: boa070270@gmail.com

    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
securityDefinitions:
    BasicAuth:
        type: basic
    ApiKeyAuth:
        type: apiKey
        in: header
        name: X-API-Key
    OAuth2:
        type: oauth2
        flow: accessCode
        authorizationUrl: https://example.com/oauth/authorize
        tokenUrl: https://example.com/oauth/token
        scopes:
            writer: Grants write access
            admin: Grants read and write access to administrative information
...
basePath: /api/v1
schemes:
    - http
    - https
parameters:
    X-Client-ID:
        in: header
        name: x-client-id
        type: string
paths:
...
    /vote/{resId}:
        parameters:
            - $ref: '#/parameters/X-Client-ID'
        x-swagger-router-controller: votes
        get:
            summary: Get votes
            operationId: getVotes
            x-query: >
                with data as (select
                cast(:resOd as varchar) resId,
                cast(:x-client-id as uuid) clientId
                ), me as (
                     select vote_id, vote from votes, data where res_id = resId and browser_id = clientId
                 ), gr as (
                     select vote_id, vote, count(*) as cnt from votes, data where res_id = resId
                     group by vote_id, vote
                )
                select 'me' as who, resId as "resId", vote_id as "voteId", vote, 1 as "count" from me, data
                union
                select 'all' as who, resId as "resId", vote_id as "voteId", vote, cnt as "count"  from gr, data
            parameters:
                -     in: path
                      name: resId
                      required: true
                      type: string
                      description: resource id
            tags:
                - public
            produces:
                - application/json
            responses:
                200:
                    x-payload: 'data'
                    description: all users
                    schema:
                        $ref: '#/definitions/VotesResponse'
                400:
                    description: Bad request
                    schema:
                        $ref: '#/definitions/StatusType'
                500:
                    description: Internal server error
                    schema:
                        $ref: '#/definitions/StatusType'
        post:
            summary: Vote
            operationId: vote
            x-query: >
                with data as (select
                cast(:resId as varchar) resId,
                cast(:x-client-id as uuid) clientId,
                cast(:voteId as int) voteId,
                cast(:value as varchar) v
                )
                insert into votes(res_id, browser_id, vote_id, vote)
                select resId, clientId, voteId, v from data
                where not exists(select * from votes where res_id = resId and browser_id = clientId and vote_id = voteId)
            tags:
                - public
            produces:
                - application/json
            parameters:
                -     in: path
                      name: resId
                      required: true
                      type: string
                      description: resource id
                -   in: body
                    name: body
                    description: vote
                    required: true
                    schema:
                        type: object
                        required:
                            - vote
                        properties:
                            vote:
                                type: string
                                description: the value of vote
                            voteId:
                                type: number
                                description: number of vote if resource has more than one
            responses:
                200:
                    x-payload: 'data.id'
                    description: Resulted id
                    schema:
                        $ref: '#/definitions/IdResponse'
                400:
                    description: Bad request
                    schema:
                        $ref: '#/definitions/StatusType'
                500:
                    description: Internal server error
                    schema:
                        $ref: '#/definitions/StatusType'
    /comment/{resId}:
        parameters:
            - $ref: '#/parameters/X-Client-ID'
        x-swagger-router-controller: votes
        get:
            summary: Get comments
            operationId: getComments
            x-query: >
                with recursive rcomments as (
                select
                id, created, res_id as "resId", nick_name as "nickName", comm_id as "commentId", comment
                from comments where res_id = :resId and coalesce(comm_id, 'unknown') = coalesce(:commentId, comm_id, 'unknown')
                union
                select
                c.id, c.created, c.res_id as "resId", c.nick_name as "nickName", c.comm_id as "commentId", c.comment
                from comments c
                inner join rcomments r on c.comm_id = r.id
                ), vote_me as (
                select r.id, vote from votes, rcomments r where res_id = r.id and browser_id = :clientId
                ), vote_gr as (
                select r.id, vote, count(*) as cnt from votes, rcomments r where res_id = r.id
                group by r.id, vote
                ), n as (
                select "commentId", count(*) as "numberOf" from rcomments group by "commentId"
                )
                select r.*, n."numberOf", vote_gr.vote vote_gr_vote, vote_gr.cnt vote_gr_count, vote_me.vote vote_me from rcomments r
                left join vote_gr on vote_gr.id = r.id
                left join vote_me on vote_me.id = r.id
                left join n n on n."commentId" = r.id
                order by r."commentId" asc NULLS FIRST, r.created
                OFFSET :from ROWS FETCH NEXT :size ROWS ONLY
            parameters:
                -     in: path
                      name: resId
                      required: true
                      type: string
                      description: resource id
                -   in: query
                    required: false
                    name: commentId
                    type: string
                -   in: query
                    required: false
                    name: size
                    type: integer
                    default: 20
                -   in: query
                    type: integer
                    name: from
                    default: 0
                    required: false
            tags:
                - public
            produces:
                - application/json
            responses:
                200:
                    x-payload: 'data'
                    description: all users
                    schema:
                        $ref: '#/definitions/CommentsResponse'
                400:
                    description: Bad request
                    schema:
                        $ref: '#/definitions/StatusType'
                500:
                    description: Internal server error
                    schema:
                        $ref: '#/definitions/StatusType'
        post:
            summary: Add comment
            operationId: addComment
            x-query: >
                with data as (select
                cast(:resId as varchar) resId,
                cast(:x-client-id as uuid) clientId,
                cast(:comment as varchar) txt,
                cast(:nickName as varchar) nick,
                cast(:commentId as varchar) commentId,
                cast(:newId as varchar) newId
                )
                insert into comments(id, res_id, browser_id, comm_id, comment, nick_name)
                select newId, resId, clientId, commentId, txt, nick from data
                returning id
            tags:
                - public
            produces:
                - application/json
            parameters:
                -     in: path
                      name: resId
                      required: true
                      type: string
                      description: resource id
                -   in: body
                    name: body
                    description: body
                    required: true
                    schema:
                        type: object
                        required:
                            - comment
                        properties:
                            comment:
                                type: string
                                description: the text of comment
                            nickName:
                                type: string
                            commentId:
                                type: string
            responses:
                200:
                    x-payload: 'data.id'
                    description: Resulted id
                    schema:
                        $ref: '#/definitions/IdResponse'
                400:
                    description: Bad request
                    schema:
                        $ref: '#/definitions/StatusType'
                500:
                    description: Internal server error
                    schema:
                        $ref: '#/definitions/StatusType'
        put:
            summary: Update comment
            operationId: updateComment
            x-query: update comments set comment = :comment where id = :commentId and browser_id = :x-client-id
            tags:
                - public
            produces:
                - application/json
            parameters:
                -     in: path
                      name: resId
                      required: true
                      type: string
                      description: comment id
                -   in: body
                    name: body
                    description: body
                    required: true
                    schema:
                        type: object
                        required:
                            - comment
                            - commentId
                        properties:
                            comment:
                                type: string
                                description: the text of comment
                            commentId:
                                type: string
                                description: id od comment
            responses:
                200:
                    x-payload: 'data.id'
                    description: Resulted id
                    schema:
                        $ref: '#/definitions/IdResponse'
                400:
                    description: Bad request
                    schema:
                        $ref: '#/definitions/StatusType'
                500:
                    description: Internal server error
                    schema:
                        $ref: '#/definitions/StatusType'
definitions:
    StatusType:
        type: object
        required:
            - code
            - message
        properties:
            code:
                type: integer
                example: 0
            details:
                type: object
            message:
                type: string
                example: successful
...
    IdResponse:
        properties:
            data:
                type: object
                properties:
                    id:
                        type: string
            status:
                $ref: '#/definitions/StatusType'
...
    VoteType:
        type: object
        properties:
            who:
                type: string
            resId:
                type: string
            voteId:
                type: number
            vote:
                type: string
            count:
                type: number
    VotesResponse:
        type: object
        properties:
            data:
                type: array
                items:
                    $ref: '#/definitions/VoteType'
            status:
                $ref: '#/definitions/StatusType'
    CommentType:
        type: object
        properties:
            id:
                type: string
            created:
                type: string
            resId:
                type: string
            nickName:
                type: string
            commentId:
                type: string
            comment:
                type: string
            isMy:
                type: boolean
            vote:
                type: array
                items:
                    $ref: '#/definitions/VoteType'
            numberOf:
                type: integer
    CommentsResponse:
        type: object
        properties:
            data:
                type: object
                properties:
                    lastComment:
                        type: string
                    numberOf:
                        type: integer
                    responses:
                        type: array
                        items:
                            $ref: '#/definitions/CommentType'
            status:
                $ref: '#/definitions/StatusType'